refactor: 將業務邏輯整合至 handler 中
