+++ src/api/socket/events/memberApplication/memberApplication.handler.ts
@@ -17,34 +17,57 @@ import {
 // Middleware
 import DataValidator from '@/middleware/data.validator';
 
-// Services
-import {
-  CreateMemberApplicationService,
-  UpdateMemberApplicationService,
-  DeleteMemberApplicationService,
-} from '@/api/socket/events/memberApplication/memberApplication.service';
+// Database
+import { database } from '@/index';
 
 export class CreateMemberApplicationHandler extends SocketHandler {
   async handle(data: any) {
     try {
       const operatorId = this.socket.data.userId;
 
-      const { userId, serverId, memberApplication } = await new DataValidator(
+      const {
+        userId,
+        serverId,
+        memberApplication: preset,
+      } = await new DataValidator(
         CreateMemberApplicationSchema,
         'CREATEMEMBERAPPLICATION',
       ).validate(data);
 
-      const { serverMemberApplicationAdd } =
-        await new CreateMemberApplicationService(
-          operatorId,
-          userId,
-          serverId,
-          memberApplication,
-        ).use();
+      const operatorMember = await database.get.member(operatorId, serverId);
+
+      if (operatorId !== userId) {
+        throw new StandardizedError({
+          name: 'ValidationError',
+          message: '無法創建非自己的會員申請',
+          part: 'CREATEMEMBERAPPLICATION',
+          tag: 'PERMISSION_DENIED',
+          statusCode: 403,
+        });
+      } else {
+        if (operatorMember && operatorMember.permissionLevel !== 1) {
+          throw new StandardizedError({
+            name: 'ValidationError',
+            message: '非遊客無法創建會員申請',
+            part: 'CREATEMEMBERAPPLICATION',
+            tag: 'PERMISSION_DENIED',
+            statusCode: 403,
+          });
+        }
+      }
+
+      // Create member application
+      await database.set.memberApplication(userId, serverId, {
+        ...preset,
+        createdAt: Date.now(),
+      });
 
       this.io
         .to(`server_${serverId}`)
-        .emit('serverMemberApplicationAdd', serverMemberApplicationAdd);
+        .emit(
+          'serverMemberApplicationAdd',
+          await database.get.serverMemberApplication(serverId, userId),
+        );
     } catch (error: any) {
       if (!(error instanceof StandardizedError)) {
         error = new StandardizedError({
@@ -67,26 +90,35 @@ export class UpdateMemberApplicationHandler extends SocketHandler {
     try {
       const operatorId = this.socket.data.userId;
 
-      const { userId, serverId, memberApplication } = await new DataValidator(
+      const {
+        userId,
+        serverId,
+        memberApplication: update,
+      } = await new DataValidator(
         UpdateMemberApplicationSchema,
         'UPDATEMEMBERAPPLICATION',
       ).validate(data);
 
-      await new UpdateMemberApplicationService(
-        operatorId,
-        userId,
-        serverId,
-        memberApplication,
-      ).use();
+      const operatorMember = await database.get.member(operatorId, serverId);
+
+      if (operatorId !== userId) {
+        if (operatorMember.permissionLevel < 5) {
+          throw new StandardizedError({
+            name: 'ValidationError',
+            message: '你沒有足夠的權限更新其他成員的會員申請',
+            part: 'UPDATEMEMBERAPPLICATION',
+            tag: 'PERMISSION_DENIED',
+            statusCode: 403,
+          });
+        }
+      }
+
+      // Update member application
+      await database.set.memberApplication(userId, serverId, update);
 
       this.io
         .to(`server_${serverId}`)
-        .emit(
-          'serverMemberApplicationUpdate',
-          userId,
-          serverId,
-          memberApplication,
-        );
+        .emit('serverMemberApplicationUpdate', userId, serverId, update);
     } catch (error: any) {
       if (!(error instanceof StandardizedError)) {
         error = new StandardizedError({
@@ -114,11 +146,22 @@ export class DeleteMemberApplicationHandler extends SocketHandler {
         'DELETEMEMBERAPPLICATION',
       ).validate(data);
 
-      await new DeleteMemberApplicationService(
-        operatorId,
-        userId,
-        serverId,
-      ).use();
+      const operatorMember = await database.get.member(operatorId, serverId);
+
+      if (operatorId !== userId) {
+        if (operatorMember.permissionLevel < 5) {
+          throw new StandardizedError({
+            name: 'ValidationError',
+            message: '你沒有足夠的權限刪除其他成員的會員申請',
+            part: 'DELETEMEMBERAPPLICATION',
+            tag: 'PERMISSION_DENIED',
+            statusCode: 403,
+          });
+        }
+      }
+
+      // Delete member application
+      await database.delete.memberApplication(userId, serverId);
 
       this.io
         .to(`server_${serverId}`)
